services/image_scrape_docker/python_requirements.txt
pandas==2.1.4
numpy==1.26.3
google-cloud-storage==2.14.0
pyarrow